update: user,comment,review

+++ src/main/java/xyz/tomorrowlearncamp/outsourcing/domain/review/service/UserWriteReviewService.java
@@ -4,11 +4,18 @@ import jakarta.validation.constraints.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
 import org.springframework.validation.annotation.Validated;
+import xyz.tomorrowlearncamp.outsourcing.domain.comment.service.OwnerWriteCommentService;
+import xyz.tomorrowlearncamp.outsourcing.domain.order.entity.UserOrderEntity;
+import xyz.tomorrowlearncamp.outsourcing.domain.order.service.UserOrderService;
 import xyz.tomorrowlearncamp.outsourcing.domain.review.entity.ReviewEntity;
 import xyz.tomorrowlearncamp.outsourcing.domain.review.enums.ErrorReviewMessage;
 import xyz.tomorrowlearncamp.outsourcing.domain.review.repository.ReviewRepository;
+import xyz.tomorrowlearncamp.outsourcing.domain.user.entity.UserEntity;
+import xyz.tomorrowlearncamp.outsourcing.domain.user.service.UserService;
 import xyz.tomorrowlearncamp.outsourcing.global.exception.InvalidRequestException;
+import xyz.tomorrowlearncamp.outsourcing.global.exception.UnauthorizedRequestException;
 
 @Service
 @RequiredArgsConstructor
@@ -17,24 +24,24 @@ public class UserWriteReviewService {
 
     private final ReviewRepository reviewRepository;
 
-//    private final UserReadService userReadService;
+    private final UserOrderService orderService;
+
+    private final UserService userService;
 
     @Transactional
     public void saveReview(
-            /*@NotBlack @Positive Long userId,
-            @NotBlank @Positive Long orderId,*/
-            @NotNull @Positive Long storeId,
+            Long userId,
+            @NotNull @Positive Long orderId,
             @NotNull @Size(min = 10, max = 100) String contents,
             @NotNull @Min(1) @Max(5) Integer star,
             String reviewImageUrl
     ) {
-//        UserEntity user = userReadService.getUser(userId);
-//        OrderEntity order = orderReadService.getOrder(orderId);
+        UserEntity user = userService.getUserEntity(userId);
+        UserOrderEntity order = orderService.getOrderById(orderId);
 
         ReviewEntity review = ReviewEntity.builder()
-//                .user(user)
-//                .order(order)
-                .storeId(storeId)
+                .user(user)
+                .orderEntity(order)
                 .contents(contents)
                 .star(star)
                 .reviewImageUrl(reviewImageUrl)
@@ -44,10 +51,14 @@ public class UserWriteReviewService {
     }
 
     @Transactional
-    public void deleteReview(@NotBlank @Positive Long reviewId) {
+    public void deleteReview(Long userId, @NotNull @Positive Long reviewId) {
         ReviewEntity deleteReview = reviewRepository.findById(reviewId).orElseThrow(
                 () -> new InvalidRequestException(ErrorReviewMessage.NOT_FOUND_REVIEW.getErrorMessage())
         );
+
+        if (ObjectUtils.nullSafeEquals(deleteReview.getUser().getId(), userId)) {
+            throw new UnauthorizedRequestException(ErrorReviewMessage.UNAUTHORIZED.getErrorMessage());
+        }
 
         reviewRepository.delete(deleteReview);
